Update Nuke and Tests

+++ src/FilamentCalcWeb/Models/Settings.cs
@@ -18,7 +18,7 @@ namespace FilamentCalculator.Models
         [DefaultValue(0.10)]
         public decimal PrinterDepricationKostsPerHour { get; set; }
         
-        public int PrinterEnergyUsageW { get; set; }
+        public decimal PrinterEnergyUsageW { get; set; }
         
         
     }
